Working on http statuses.

+++ src/main/java/pl/kosiorski/controller/TaskController.java
@@ -1,19 +1,20 @@
 package pl.kosiorski.controller;
 
 import org.hibernate.ObjectNotFoundException;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 import pl.kosiorski.dto.TaskDto;
 import pl.kosiorski.exception.NoAuthenticationException;
+import pl.kosiorski.exception.NoAuthorizationException;
 import pl.kosiorski.service.AuthService;
 import pl.kosiorski.service.TaskService;
 import pl.kosiorski.service.UserService;
 
 import javax.validation.Valid;
 import java.util.List;
+import java.util.NoSuchElementException;
 
 @RestController
 @RequestMapping("/tasks")
@@ -71,22 +72,23 @@ public class TaskController {
     return null;
   }
 
-  // TODO block the possibility of deleting tasks of other users
-
   @DeleteMapping("/{id}")
   public ResponseEntity delete(@PathVariable Long id, @RequestHeader(HEADER_KEY) String token) {
     try {
-      if (authService.validateToken(token)) {
+      if (authService.validateToken(token) && taskService.taskBelongToUser(token, id)) {
         taskService.deleteById(id);
         return new ResponseEntity(HttpStatus.OK);
       }
     } catch (NoAuthenticationException e) {
       throw new ResponseStatusException(HttpStatus.FORBIDDEN, e.getMessage(), e);
 
-    } catch (EmptyResultDataAccessException e) {
-      throw new ResponseStatusException(
-          HttpStatus.BAD_REQUEST, "Task with the id: " + id + " does not exist");
+    } catch (NoSuchElementException e) {
+      throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Task with the id: " + id + " does not exist");
+
+    } catch (NoAuthorizationException e) {
+      throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, e.getMessage(), e);
     }
+
     return null;
   }
